added write to firebase realtime db
created basic screen to enter text and submit it to the mock firebase db. this requires you to sign in first with email and password (or sign up if never logged in before), then you can enter in a short message and have it saved to firebase. you can either close app and stay signed in or click the more button at the top right and sign out. this will ask you to sign in when you try to make another post.

**to do:**
* work on displaying results in the same or another activity.
* figure out why it jumps back to the main activity after making a post.