added ability to read from firebase db
added recyclerview that reads all "posts" in my test firebase db. it updates in real time when you make a post.

issue - every time the floating action button is pressed to submit a post, the activity closes and goes back to MainActivity, but it does still update with the post (which is visible when you go back into the activity)