retooled reply display, added listeners
restructured how the replies window looks by adding logic for differentiating between sender (you) and receiver (the other guy). this is important for karma points because it will serve as a way to block repliers from giving karma points to each other. only the requestor should be able to do this.

added click and touch listeners (basically same thing); this will allow for tapping on an icon to send another user karma points.