added date/time stamp
changed how timestamp is generated - was being created by firebase, now created on device (but it is still stored in firebase). this allows the models to pull down the timestamp from database, whereas before timestamp was set as a result of pushing to the database.

the timestamp is in a unix time format. when displayed, it is converted to MM/DD/YYYY HH:MM:SS A

also removed the author name (which was just showing uid) and put timestamp in its place