changed carousel from viewPager to fragment
made the carousel on the maps activity screen a fragment instead of a viewPager. this has no real effect on things, except helped to remove a few lines of code

removed some more code that was leftover stuff

also added a dummy_mapview_fragment.xml file that holds a static image of google maps. this is in turn displayed in android studio on the activity_maps.xml view for our use only. now if you open activity_maps.xml, you will see a nice rendering of what this view looks like in the app (all this purely cosmetic)