login and posting fix
both the anonymous login and posting appear to be fixed, at least for my device. only 1 userid is created, it is written to the db after creation. on first run, it goes through the intro and then drops you into the map. when making a request, it does not give the no user error and it moves back to the maps screen, not intro nor does the app close. more testing on other devices may be required.